Split implementation of SimpleCache from interface RespCache.
