+++ source/extensions/filters/http/ext_authz/BUILD
@@ -17,7 +17,6 @@ envoy_cc_library(
     srcs = ["ext_authz.cc"],
     hdrs = [
         "ext_authz.h",
-        "//source/common/resp_cache:resp_cache.h",
     ],
     deps = [
         "//envoy/http:codes_interface",
@@ -30,6 +29,7 @@ envoy_cc_library(
         "//source/common/common:minimal_logger_lib",
         "//source/common/http:codes_lib",
         "//source/common/http:utility_lib",
+        "//source/common/resp_cache",
         "//source/common/router:config_lib",
         "//source/common/runtime:runtime_protos_lib",
         "//source/extensions/filters/common/ext_authz:ext_authz_grpc_lib",
